+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.2] - 2023-01-16
+
+### Removed
+
+- Requirement for `Debug` and `Clone` for types `T` in `SearchResults`
+  is no longer required in order for `Client#search` to return results.
+  Now if the type `T` **does** implement both/either trait then
+  `SearchResults` will also have that functionality.
+
 ## [0.1.1] - 2023-01-15
 
 ### Changed